Google, Facebook, Amazon
Given an array of integers A of size N. A represents a histogram i.e A[i] denotes height of 

the ith histogram’s bar. Width of each bar is 1.
Given an array of integers A of size N. A represents a histogram i.e A[i] denotes height of 

the ith histogram’s bar. Width of each bar is 1.
The largest rectangle is shown in the shaded area, which has area = 10 unit.

Find the area of largest rectangle in the histogram